requirements.txt
faker>=22.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit
langchain
langchain-groq
langchain-community
psycopg2-binary
python-dotenv
plotly
pandas